+++ src/Ext/TechnoType/Body.cpp
@@ -11,6 +11,7 @@
 #include <Ext/BuildingType/Body.h>
 #include <Ext/BulletType/Body.h>
 #include <Ext/Techno/Body.h>
+#include <Ext/House/Body.h>
 #include <New/Type/InsigniaTypeClass.h>
 
 #include <Utilities/GeneralUtils.h>
@@ -1328,6 +1329,138 @@ void TechnoTypeExt::ExtData::LoadFromINIFile(CCINIClass* const pINI)
 		this->InterceptorType.reset();
 	}
 
+	this->ConsideredNaval.Read(exINI, pSection, "ConsideredNaval");
+	this->ConsideredVehicle.Read(exINI, pSection, "ConsideredVehicle");
+	this->ConsideredSecretLabTech.Read(exINI, pSection, "ConsideredSecretLabTech");
+
+	// Secret.RequiredHouses contains a list of HouseTypeClass indexes
+	if (pINI->ReadString(pSection, "SecretLab.RequiredHouses", "", Phobos::readBuffer) > 0)
+	{
+		char* context = nullptr;
+		this->Secret_RequiredHouses.clear();
+
+		for (char* cur = strtok_s(Phobos::readBuffer, Phobos::readDelims, &context); cur; cur = strtok_s(nullptr, Phobos::readDelims, &context))
+		{
+			std::string item(cur);
+			this->Secret_RequiredHouses.push_back(item);
+		}
+	}
+
+	// Secret.ForbiddenHouses contains a list of HouseTypeClass indexes
+	if (pINI->ReadString(pSection, "SecretLab.ForbiddenHouses", "", Phobos::readBuffer) > 0)
+	{
+		char* context = nullptr;
+		this->Secret_ForbiddenHouses.clear();
+
+		for (char* cur = strtok_s(Phobos::readBuffer, Phobos::readDelims, &context); cur; cur = strtok_s(nullptr, Phobos::readDelims, &context))
+		{
+			std::string item(cur);
+			this->Secret_ForbiddenHouses.push_back(item);
+		}
+	}
+
+	// Prerequisite.RequiredTheaters contains a list of theader names
+	if (pINI->ReadString(pSection, "Prerequisite.RequiredTheaters", "", Phobos::readBuffer) > 0)
+	{
+		char* context = nullptr;
+		this->Prerequisite_RequiredTheaters.clear();
+
+		for (char* cur = strtok_s(Phobos::readBuffer, Phobos::readDelims, &context); cur; cur = strtok_s(nullptr, Phobos::readDelims, &context))
+		{
+			int index = Theater::FindIndex(cur);
+
+			if (index != -1)
+				this->Prerequisite_RequiredTheaters.push_back(index);
+		}
+	}
+
+	// Prerequisite with Generic Prerequistes support.
+	// Note: I have no idea of what could happen in all the game engine logics if I push the negative indexes of the Ares generic prerequisites directly into the original Prerequisite tag... for that reason this tag is duplicated for working with it
+	if (pINI->ReadString(pSection, "Prerequisite", "", Phobos::readBuffer) > 0)
+	{
+		char* context = nullptr;
+		this->Prerequisite.clear();
+
+		for (char* cur = strtok_s(Phobos::readBuffer, Phobos::readDelims, &context); cur; cur = strtok_s(nullptr, Phobos::readDelims, &context))
+		{
+			int idx = TechnoTypeClass::FindIndex(cur);
+
+			if (idx >= 0)
+			{
+				this->Prerequisite.push_back(idx);
+			}
+			else
+			{
+				int index = HouseExt::FindGenericPrerequisite(cur);
+
+				if (index < 0)
+					this->Prerequisite.push_back(index);
+			}
+		}
+	}
+
+	// Prerequisite.Negative with Generic Prerequistes support
+	if (pINI->ReadString(pSection, "Prerequisite.Negative", "", Phobos::readBuffer) > 0)
+	{
+		char* context = nullptr;
+		this->Prerequisite_Negative.clear();
+
+		for (char* cur = strtok_s(Phobos::readBuffer, Phobos::readDelims, &context); cur; cur = strtok_s(nullptr, Phobos::readDelims, &context))
+		{
+			int idx = TechnoTypeClass::FindIndex(cur);
+
+			if (idx >= 0)
+			{
+				this->Prerequisite_Negative.push_back(idx);
+			}
+			else
+			{
+				int index = HouseExt::FindGenericPrerequisite(cur);
+
+				if (index < 0)
+					this->Prerequisite_Negative.push_back(index);
+			}
+		}
+	}
+
+	// Prerequisite.ListX with Generic Prerequistes support
+	this->Prerequisite_Lists.Read(exINI, pSection, "Prerequisite.Lists");
+
+	if (this->Prerequisite_Lists.Get() >= 0)
+	{
+		this->Prerequisite_ListVector.clear();
+
+		for (int i = 1; i <= Prerequisite_Lists.Get(); i++)
+		{
+			char keySection[32];
+			_snprintf_s(keySection, sizeof(keySection), "Prerequisite.List%d", i);
+
+			DynamicVectorClass<int> objectsList;
+			char* context = nullptr;
+			pINI->ReadString(pSection, keySection, "", Phobos::readBuffer);
+
+			for (char* cur = strtok_s(Phobos::readBuffer, Phobos::readDelims, &context); cur; cur = strtok_s(nullptr, Phobos::readDelims, &context))
+			{
+				int idx = TechnoTypeClass::FindIndex(cur);
+
+				if (idx >= 0)
+				{
+					objectsList.AddItem(idx);
+				}
+				else
+				{
+					int index = HouseExt::FindGenericPrerequisite(cur);
+
+					if (index < 0)
+						objectsList.AddItem(index);
+				}
+			}
+
+			this->Prerequisite_ListVector.push_back(objectsList);
+			objectsList.Clear();
+		}
+	}
+
 	if (pThis->WhatAmI() != AbstractType::BuildingType)
 	{
 		if (this->DroppodType == nullptr)
@@ -1569,6 +1702,17 @@ void TechnoTypeExt::ExtData::Serialize(T& Stm)
 
 		.Process(this->Explodes_KillPassengers)
 		.Process(this->Explodes_DuringBuildup)
+		.Process(this->Prerequisite_RequiredTheaters)
+		.Process(this->Prerequisite)
+		.Process(this->Prerequisite_Negative)
+		.Process(this->Prerequisite_Lists)
+		.Process(this->Prerequisite_ListVector)
+		.Process(this->ConsideredNaval)
+		.Process(this->ConsideredVehicle)
+		.Process(this->ConsideredSecretLabTech)
+		.Process(this->Secret_RequiredHouses)
+		.Process(this->Secret_ForbiddenHouses)
+
 		.Process(this->DeployFireWeapon)
 		.Process(this->TargetZoneScanType)